Replace firmware restoration instructions with link to duplicate Help Center content
Unfortunately, the firmware packaging system includes the version number in the firmware binary filename. The exact
binary filename is referenced in the firmware restoration instructions, which means that they must be updated on every
release or else become obsolete.

Equivalent instructions have also been published in the Arduino Help Center. The fact that the maintainers of both
projects have failed to keep their instructions updated through the course of multiple releases and over 6 months shows
that is is not feasible to maintain redundant copies of the instructions.

Since these instructions are not directly relevant to the updater scripts, they are not really in scope for the script's
readme. The Help Center article is significantly discoverable to the user, slightly higher quality that the variant
here, and it makes more sense to put the responsibility for such content on the maintainers of the Help Center. For
these reasons, the redundant instructions are removed from this repository, replaced with a link to the Help Center
article.