Update Help Center article links
The project documentation contains links to an Arduino Help Center article that provides instructions for flashing the
firmware by using the "espflash" tool directly.

At the time those links were added, the instructions were in a subsection of an article that covered all methods of
updating the firmware. Since that time, the instructions for using the "espflash" tool directly were moved to a
dedicated article.

A link was added from the previous article to the new article, so it is still possible for the reader to find the
intended content even via the previous link. However, that is inconvenient and prone to user error. So it is best to
link directly to the intended content.